pyproject.toml
[tool.poetry]
name = "drf-project"
version = "0.1.0"
description = ""
authors = ["Makoto <makoto@mak.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.2"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
black = "^24.4.2"
isort = "^5.13.2"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.6.0"
pyjwt = "^2.8.0"
drf-yasg = "^1.21.7"
packaging = "^24.1"
pyyaml = "^6.0.2"
stripe = "^10.8.0"
redis = "^5.0.8"
forex-python = "^1.8"
setuptools = "^72.2.0"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
